Fix categorize layer bug where layer source paths were changed wrongly
The problem with the current approach is that relative paths are stored depending
on the path of the QGIS project. When we use QgsProject to write the project to a new location
then QGIS will update the layer sources to be relative to the new project location.
This obviously breaks layers.
The new approach makes sure the project is stored in the same directory as the old
project (which is the default behavior if no destination location is specified) and
then moves the project to the correct path with the correct name.
This makes sure no layer sources are changed and everything is kept as is.